requirements.txt
configparser==3.5.0
Django==2.0.3
djangorestframework==3.7.7
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
